Moving /atom to active endpoints
Removing old vangogh endpoints
