--- rest_vg/refresh_redux_assets.go
@@ -1,11 +1,0 @@
-package rest_vg
-
-func RefreshReduxAssets(properties ...string) (err error) {
-	if rdx, err = rdx.RefreshReader(); err != nil {
-		return err
-	}
-	if err := rdx.MustHave(properties...); err != nil {
-		return err
-	}
-	return err
-}
